Add a globals scope to ScopeManager to keep track of global vars.
When a load is triggered, the globals scope will be queried from the
ScopeManager if the load fails in the local context.

When a store is triggered, the globals scope will be stored into instead
if a global is created (e.g. by assigning to an unbound variable
locally, because Javascript is stupid about that). Closures also
transfer their contents into the global scope when their original
containing scope dies, but I'm going to change this soon to make more
sense, so that closures get their own scopes at the ScopeManager level.