Perform incremental pointer coloring.
Instead of keeping track of a set of root references per-scope, Scope no
longer keeps track of any of this information, and instead relies on the
alloc_box to keep track of live references, etc. References are assumed
to be grey when allocated, and are "condemned" (made white) when they
are overwritten. This should speed up garbage collection in the marking
phase.